gateway/requirements.txt
# Gateway dependencies
flask>=3.0
flask-cors>=4.0
httpx>=0.27
python-dotenv>=1.0
